Create Locations_Select_ByGeo
This Query was a fairly early build of how I designed out how to get records within a given mile radius. While in the end this specific proc did not end up getting used, the core logic behind the selecting by a radius would be implemented and used throughout most of the rest of the project. See Listings_SelectIds_ByGeo.